+++ outputs.tf
@@ -18,17 +18,17 @@ output "prometheus_alert_groups" {
 }
 
 output "image_name" {
-  description = "Name of the docker image used for the aws-health-status container."
+  description = "Name of the docker image used for the redis-exporter container."
   value       = var.enabled ? var.image_name : ""
 }
 
 output "image_version" {
-  description = "Tag of the docker image used for the aws-health-status container."
+  description = "Tag of the docker image used for the redis-exporter container."
   value       = var.enabled ? var.image_version : ""
 }
 
 output "image_pull_policy" {
-  description = "Image pull policy defined on the aws-health-status container."
+  description = "Image pull policy defined on the redis-exporter container."
   value       = var.enabled ? var.image_pull_policy : ""
 }